+++ src/main/java/sidic/entities/TextosFacturas.java
@@ -37,7 +37,7 @@ public class TextosFacturas implements Serializable {
     @Id
     @Basic(optional = false)
     @Column(nullable = false, precision = 22)
-    private Double codigo;
+    private Long codigo;
     @Column(length = 80)
     private String texto1;
     @Column(length = 80)
@@ -51,15 +51,15 @@ public class TextosFacturas implements Serializable {
     public TextosFacturas() {
     }
 
-    public TextosFacturas(Double codigo) {
+    public TextosFacturas(Long codigo) {
         this.codigo = codigo;
     }
 
-    public Double getCodigo() {
+    public Long getCodigo() {
         return codigo;
     }
 
-    public void setCodigo(Double codigo) {
+    public void setCodigo(Long codigo) {
         this.codigo = codigo;
     }
